+++ packages/core/src/index.ts
@@ -5,3 +5,4 @@ export { Ceramic } from './ceramic.js'
 
 export * from './store/ikv-store.js'
 export * from './ceramic.js'
+export * from './pubsub/pubsub-message.js'