+++ packages/stream-tests/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.39.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tests@2.38.0...@ceramicnetwork/stream-tests@2.39.0-rc.0) (2023-09-06)
+
+
+### Features
+
+* Report number of pending anchors in metrics and nodeStatus output ([#2918](https://github.com/ceramicnetwork/js-ceramic/issues/2918)) ([c93bb2a](https://github.com/ceramicnetwork/js-ceramic/commit/c93bb2afe6e346e160ba4911c2a94fdd380de52b))
+
+
+
+
+
 # [2.38.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tests@2.38.0-rc.1...@ceramicnetwork/stream-tests@2.38.0) (2023-08-28)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tests